Change window level from float to double
Note: Setting fractional values is currently still prohibited by the window/level slider.

+++ src/main/scala/scalismo/ui/model/properties/WindowLevel.scala
@@ -6,4 +6,4 @@ package scalismo.ui.model.properties
  * In one sentence: "Making the window wider or narrower decreases or increases the display contrast;
  * moving the level left or right changes the display brightness."
  */
-case class WindowLevel(window: Float, level: Float)
+case class WindowLevel(window: Double, level: Double)
